feat(core): improve the mechanism to change tab
We no longer scrape the DOM to get the available views and switch between them. Instead, we rely on
the native javascript objects offered by Home Assistant. This method is more reliable and allow us
to switch view even when tabs are not visible in the UI, e.g. when inside a subview.

Ref. #68